+++ docs/changelog.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+## 2026-07-06
+
+- Add `-websocket` flag to agent connection to support HTTP-only reverse proxies (traefik-k8s ingress).
+
 ## 2026-06-23
 
 - Fix deprecated `-jnlpUrl` argument by using `-url` and `-name` when connecting to Jenkins.